Reconfigure the service manager instead of aggregate config
The various plugin managers are built using configured services — but
those services are typically not yet configured at the time the
ServiceListener is invoked. As such, we need to re-configure the service
manager prior to attempts to create the various plugin managers.

Since the logic is the same between merging configuration for the
service manager as it is for plugin managers, I have extracted methods
for this that each routine can invoke.

Finally, I've also added logic to *not* create a plugin manager if we
did not receive any configuration for it.