Do not raise an exception if a given plugin manager is not defined.
The `ServiceListener::onLoadModulesPost()` listener raises an exception
if a given plugin manager instance is unavailable, and does not exist in
the application service manager. This impacts migrating the various
plugin managers to their respective components:

- If they are registered as part of the module configuration, but that
  component is not present, then an exception is raised.
- We cannot add delegator factories on the `ServiceListener` from
  component module classes, as the `ServiceListener` is retrieved
  *before* loading modules.

One solution is to assume that if the plugin manager is not present,
nothing needs to be done. This allows components to map the plugin
loader services they expose via their `Module` classes, ensuring that
once the listener triggers, it *can* access the service and configure
it.

Ideally, however, the component should also be able to notify the
`ServiceListener` of plugin managers; hopefully a later commit can
address that problem.