Added zend-config as a requirement (vs dev-only)
Per zendframework/ZendSkeletonApplication#329, the default usage pattern for
zend-modulemanager does configuration merging, which, at this time, requires
zend-config. If this is not present in zend-modulemanager, any application using
the default listeners will need to add the requirement manually in order to
work (which is precisely what happened with the aforementioned PR).