Write cache to temporary file first and then rename
Alternative solution for issue reported in #87.
The problem is that every process running process check if the cache file
exists and if it doesn't it tries to create it. In case of big
configuration creating cache file can take a while. And while it is
generated other process can read "malformed" file, which is the issue.

As locking file does not work out of the box with include/require
(see: https://bugs.php.net/bug.php?id=52895)
we can add flock around include/require (where the cache file is read)
or apply the solution from this PR.